+++ idempotencytest/memory_test.go
@@ -1,0 +1,14 @@
+package idempotencytest_test
+
+import (
+	"testing"
+
+	"github.com/bowlinedev/bowline"
+	"github.com/bowlinedev/bowline/idempotencytest"
+)
+
+func TestMemoryStore(t *testing.T) {
+	idempotencytest.Verify(t, func(t *testing.T) bowline.IdempotencyStore {
+		return bowline.MemoryIdempotencyStore()
+	})
+}